\name{katrina.bydoc}
\alias{katrina.bydoc}
\docType{data}
\title{Organizational collaboration networks in response to Hurricane Katrina (2005), organized by source document}
\description{A list of \code{\link[network]{network}} objects, each containing all of the network information extracted from a given source document.}
\format{
	A list of 187 \code{\link[network]{network}} objects:
	
	\tabular{lll}{
		\bold{List element index} \tab \bold{List element type} \tab \bold{List element name(source document ID)}\cr
		\samp{[[1]]} \tab network \tab 10005500\cr		\samp{[[2]]} \tab network \tab 10005501\cr		\samp{[[3]]} \tab network \tab 10005502\cr		\samp{[[4]]} \tab network \tab 10005503\cr		\samp{[[5]]} \tab network \tab 10005504\cr		\samp{[[6]]} \tab network \tab 10005505\cr		\samp{[[7]]} \tab network \tab 10005506\cr		\samp{[[8]]} \tab network \tab 10005507\cr		\samp{[[9]]} \tab network \tab 10005508\cr		\samp{[[10]]} \tab network \tab 10005509\cr		\samp{[[11]]} \tab network \tab 10005510\cr		\samp{[[12]]} \tab network \tab 10005511\cr		\samp{[[13]]} \tab network \tab 15000830\cr		\samp{[[14]]} \tab network \tab 15000831\cr		\samp{[[15]]} \tab network \tab 15000901\cr		\samp{[[16]]} \tab network \tab 15000902\cr		\samp{[[17]]} \tab network \tab 15000904\cr		\samp{[[18]]} \tab network \tab 15000905\cr		\samp{[[19]]} \tab network \tab 19005500\cr		\samp{[[20]]} \tab network \tab 19005501\cr		\samp{[[21]]} \tab network \tab 19005502\cr		\samp{[[22]]} \tab network \tab 19005503\cr		\samp{[[23]]} \tab network \tab 19005504\cr		\samp{[[24]]} \tab network \tab 19005505\cr		\samp{[[25]]} \tab network \tab 19005506\cr		\samp{[[26]]} \tab network \tab 19005507\cr		\samp{[[27]]} \tab network \tab 19005508\cr		\samp{[[28]]} \tab network \tab 19005509\cr		\samp{[[29]]} \tab network \tab 19005510\cr		\samp{[[30]]} \tab network \tab 19005511\cr		\samp{[[31]]} \tab network \tab 19005512\cr		\samp{[[32]]} \tab network \tab 19005513\cr		\samp{[[33]]} \tab network \tab 19005514\cr		\samp{[[34]]} \tab network \tab 19005515\cr		\samp{[[35]]} \tab network \tab 19005516\cr		\samp{[[36]]} \tab network \tab 19005517\cr		\samp{[[37]]} \tab network \tab 19005518\cr		\samp{[[38]]} \tab network \tab 19005519\cr		\samp{[[39]]} \tab network \tab 19005520\cr		\samp{[[40]]} \tab network \tab 19005521\cr		\samp{[[41]]} \tab network \tab 19005522\cr		\samp{[[42]]} \tab network \tab 19005523\cr		\samp{[[43]]} \tab network \tab 19005524\cr		\samp{[[44]]} \tab network \tab 19005525\cr		\samp{[[45]]} \tab network \tab 19005526\cr		\samp{[[46]]} \tab network \tab 19005527\cr		\samp{[[47]]} \tab network \tab 19005528\cr		\samp{[[48]]} \tab network \tab 19005529\cr		\samp{[[49]]} \tab network \tab 19005530\cr		\samp{[[50]]} \tab network \tab 19005531\cr		\samp{[[51]]} \tab network \tab 19005532\cr		\samp{[[52]]} \tab network \tab 19005533\cr		\samp{[[53]]} \tab network \tab 19005534\cr		\samp{[[54]]} \tab network \tab 19005535\cr		\samp{[[55]]} \tab network \tab 19005536\cr		\samp{[[56]]} \tab network \tab 19005537\cr		\samp{[[57]]} \tab network \tab 19455500\cr		\samp{[[58]]} \tab network \tab 19455501\cr		\samp{[[59]]} \tab network \tab 19455502\cr		\samp{[[60]]} \tab network \tab 19455503\cr		\samp{[[61]]} \tab network \tab 19455504\cr		\samp{[[62]]} \tab network \tab 19455505\cr		\samp{[[63]]} \tab network \tab 20005500\cr		\samp{[[64]]} \tab network \tab 20005501\cr		\samp{[[65]]} \tab network \tab 20005502\cr		\samp{[[66]]} \tab network \tab 20005503\cr		\samp{[[67]]} \tab network \tab 20005504\cr		\samp{[[68]]} \tab network \tab 20005505\cr		\samp{[[69]]} \tab network \tab 20005506\cr		\samp{[[70]]} \tab network \tab 20005507\cr		\samp{[[71]]} \tab network \tab 20005508\cr		\samp{[[72]]} \tab network \tab 20005509\cr		\samp{[[73]]} \tab network \tab 20005510\cr		\samp{[[74]]} \tab network \tab 20005511\cr		\samp{[[75]]} \tab network \tab 20005512\cr		\samp{[[76]]} \tab network \tab 20710904\cr		\samp{[[77]]} \tab network \tab 33005500\cr		\samp{[[78]]} \tab network \tab 33005501\cr		\samp{[[79]]} \tab network \tab 33005502\cr		\samp{[[80]]} \tab network \tab 33005503\cr		\samp{[[81]]} \tab network \tab 33005504\cr		\samp{[[82]]} \tab network \tab 33005505\cr		\samp{[[83]]} \tab network \tab 34700902\cr		\samp{[[84]]} \tab network \tab 34700903\cr		\samp{[[85]]} \tab network \tab 34700904\cr		\samp{[[86]]} \tab network \tab 35005500\cr		\samp{[[87]]} \tab network \tab 35005501\cr		\samp{[[88]]} \tab network \tab 35005502\cr		\samp{[[89]]} \tab network \tab 35005503\cr		\samp{[[90]]} \tab network \tab 35005504\cr		\samp{[[91]]} \tab network \tab 35005505\cr		\samp{[[92]]} \tab network \tab 35005506\cr		\samp{[[93]]} \tab network \tab 35005507\cr		\samp{[[94]]} \tab network \tab 35005508\cr		\samp{[[95]]} \tab network \tab 52400831\cr		\samp{[[96]]} \tab network \tab 53005500\cr		\samp{[[97]]} \tab network \tab 53005501\cr		\samp{[[98]]} \tab network \tab 53005502\cr		\samp{[[99]]} \tab network \tab 53005503\cr		\samp{[[100]]} \tab network \tab 53005504\cr		\samp{[[101]]} \tab network \tab 53005505\cr		\samp{[[102]]} \tab network \tab 53005506\cr		\samp{[[103]]} \tab network \tab 53005507\cr		\samp{[[104]]} \tab network \tab 53005512\cr		\samp{[[105]]} \tab network \tab 53005513\cr		\samp{[[106]]} \tab network \tab 53005514\cr		\samp{[[107]]} \tab network \tab 53005515\cr		\samp{[[108]]} \tab network \tab 53400902\cr		\samp{[[109]]} \tab network \tab 53415500\cr		\samp{[[110]]} \tab network \tab 53415501\cr		\samp{[[111]]} \tab network \tab 53415502\cr		\samp{[[112]]} \tab network \tab 53415504\cr		\samp{[[113]]} \tab network \tab 53415505\cr		\samp{[[114]]} \tab network \tab 53415506\cr		\samp{[[115]]} \tab network \tab 53415507\cr		\samp{[[116]]} \tab network \tab 56005500\cr		\samp{[[117]]} \tab network \tab 56005501\cr		\samp{[[118]]} \tab network \tab 56005502\cr		\samp{[[119]]} \tab network \tab 56005503\cr		\samp{[[120]]} \tab network \tab 56005504\cr		\samp{[[121]]} \tab network \tab 56005505\cr		\samp{[[122]]} \tab network \tab 88010828\cr		\samp{[[123]]} \tab network \tab 88010829\cr		\samp{[[124]]} \tab network \tab 88010830\cr		\samp{[[125]]} \tab network \tab 88010831\cr		\samp{[[126]]} \tab network \tab 88010901\cr		\samp{[[127]]} \tab network \tab 88010902\cr		\samp{[[128]]} \tab network \tab 88010903\cr		\samp{[[129]]} \tab network \tab 88010904\cr		\samp{[[130]]} \tab network \tab 88010905\cr		\samp{[[131]]} \tab network \tab 88015500\cr		\samp{[[132]]} \tab network \tab 88015501\cr		\samp{[[133]]} \tab network \tab 88020829\cr		\samp{[[134]]} \tab network \tab 88020901\cr		\samp{[[135]]} \tab network \tab 88020902\cr		\samp{[[136]]} \tab network \tab 88030904\cr		\samp{[[137]]} \tab network \tab 88030905\cr		\samp{[[138]]} \tab network \tab 99050901\cr		\samp{[[139]]} \tab network \tab 99050902\cr		\samp{[[140]]} \tab network \tab 99050903\cr		\samp{[[141]]} \tab network \tab 99050904\cr		\samp{[[142]]} \tab network \tab 99050905\cr		\samp{[[143]]} \tab network \tab 53415504\cr		\samp{[[144]]} \tab network \tab 53415505\cr		\samp{[[145]]} \tab network \tab 53415506\cr		\samp{[[146]]} \tab network \tab 53415507\cr		\samp{[[147]]} \tab network \tab 56005500\cr		\samp{[[148]]} \tab network \tab 56005501\cr		\samp{[[149]]} \tab network \tab 56005502\cr		\samp{[[150]]} \tab network \tab 56005503\cr		\samp{[[151]]} \tab network \tab 56005504\cr		\samp{[[152]]} \tab network \tab 56005505\cr		\samp{[[153]]} \tab network \tab 88010828\cr		\samp{[[154]]} \tab network \tab 88010829\cr		\samp{[[155]]} \tab network \tab 88010830\cr		\samp{[[156]]} \tab network \tab 88010831\cr		\samp{[[157]]} \tab network \tab 88010901\cr		\samp{[[158]]} \tab network \tab 88010902\cr		\samp{[[159]]} \tab network \tab 88010903\cr		\samp{[[160]]} \tab network \tab 88010904\cr		\samp{[[161]]} \tab network \tab 88010905\cr		\samp{[[162]]} \tab network \tab 88015500\cr		\samp{[[163]]} \tab network \tab 88015501\cr		\samp{[[164]]} \tab network \tab 88020829\cr		\samp{[[165]]} \tab network \tab 88020901\cr		\samp{[[166]]} \tab network \tab 88020902\cr		\samp{[[167]]} \tab network \tab 88030904\cr		\samp{[[168]]} \tab network \tab 88030905\cr		\samp{[[169]]} \tab network \tab 99050901\cr		\samp{[[170]]} \tab network \tab 99050902\cr		\samp{[[171]]} \tab network \tab 99050903\cr		\samp{[[172]]} \tab network \tab 99050904\cr		\samp{[[173]]} \tab network \tab 99050905\cr		\samp{[[174]]} \tab network \tab 99060903\cr		\samp{[[175]]} \tab network \tab 99060904\cr		\samp{[[176]]} \tab network \tab 99060905\cr		\samp{[[177]]} \tab network \tab 99070825\cr		\samp{[[178]]} \tab network \tab 99070826\cr		\samp{[[179]]} \tab network \tab 99070829\cr		\samp{[[180]]} \tab network \tab 99070830\cr		\samp{[[181]]} \tab network \tab 99070901\cr		\samp{[[182]]} \tab network \tab 99070902\cr		\samp{[[183]]} \tab network \tab 99070903\cr		\samp{[[184]]} \tab network \tab 99070904\cr		\samp{[[185]]} \tab network \tab 99070905\cr		\samp{[[186]]} \tab network \tab 99085500\cr		\samp{[[187]]} \tab network \tab 99085501
	}
}
\details{
	Each network in katrina.bydoc has the following unique network attributes:

	\tabular{lll}{
		\bold{Attribute name} \tab \bold{Attribute name} \tab \bold{Description}\cr
		Doc.ID \tab character \tab The document idenfication number (eight digits) of the document supplying the network information\cr
		Source.ID \tab numeric \tab The source organiztion identificaiton number (four digits) of the source organization that released this document\cr
		Source.Name \tab character \tab The source organization name (short format) of the source organization that released this document\cr
		Posted.Date \tab numeric \tab The posting date (in DDMMYY format) of this document\cr
		Start.Date \tab numeric \tab The start date (in DDMMYY format) of this document\cr
		End.Date \tab numeric \tab The end date (in DDMMYY format) of this document
	}
}
\author{Carter T. Butts \email{buttsc@uci.edu}, Ryan M. Acton \email{rmacton@gmail.com}, and Christopher Marcum \email{christopher.steven.marcum@gmail.com}.}
\keyword{data}
\source{For source reference information, see \code{\link[katrina]{katrina-package}}.}
\seealso{\code{\link[katrina]{katrina-package}}, \code{\link[katrina]{katrina.bydate}}, \code{\link[katrina]{katrina.bysrc}}, \code{\link[katrina]{katrina.combined}},\code{\link[katrina]{katrina.lineage}}.}
\examples{
    data(katrina.bydoc)	# Load in the data

    # For each document from which the data were extracted,
    # tell me the number of edges in each respective
    # network
    for(i in 1:length(katrina.bydoc))
        cat("katrina.bydoc[[",i,"]]: ",
        network.edgecount(katrina.bydoc[[i]]),"\n",sep="")
}
